fixup on visual prompt w/ 0 boxes: remove all detections in this case
when feeding "visual" + 0 boxes, the model will sometimes just detect the most salient objects in the scene (while we want it to output nothing since it receives 0 visual box prompts and no valid text prompt, during a VG propagation usage). In this PR, we add a special parameter ("allow_new_detections") to handle this case.