fixup: directly fold presence logits into the model's logits predictions in the video predictor
In this PR, we fix an inference error in the current video model with presence scores. Unlike the image grounding model where the presence score is handled in as a part of the post processing, in the video model, the presence score needs to be in the detection logits since it would be consumed by other parts of the model.